\name{IFSPlot-package}
\alias{IFSPlot-package}
\alias{IFSPlot}
\docType{package}
\title{
  Plot fractals using the Iterated Function Systems in C++.
}
\description{
  Plot fractals using the Iterated Function Systems in C++.
}
\details{
  Fractals such as clouds or plants can be generated using the iteration of one or more affine transformations.
}
\author{
Cristian Sandoval-Pineda.

Maintainer: Cristian Sandoval-Pineda <sandovalp.ce@gmail.com>
}
\keyword{ package }
