% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{RunIFS}
\alias{RunIFS}
\title{RunChaosGame}
\usage{
RunIFS(transformation, translation, probability, iterations, width, height,
  set_palette = 0)
}
\arguments{
\item{transformation}{A list of size n, the 2 x 2 transformation matrices.}

\item{translation}{A list of size n, the 2 x 2 translational vectors.}

\item{probability}{A list of size n, the weighted probabilities.}

\item{iterations}{The number of iterations/points.}

\item{width}{The width of the image.}

\item{height}{The height of the image.}

\item{set_palette}{0 = default palette; 1 = grayscale; 2 = rich.colors.}
}
\value{
A matrix of pixels.
}
\description{
This function allows you to plot some fractals using the Iterated Function Systems theory.
}
