% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{RunIFS3D}
\alias{RunIFS3D}
\title{RunIFS3D}
\usage{
RunIFS3D(transformation, translation, probability, iterations, pixels, width,
  height, set_palette = 0)
}
\arguments{
\item{transformation}{A list of size n, the 3 x 3 transformation matrices.}

\item{translation}{A list of size n, the 3 x 1 translational vectors.}

\item{probability}{A list of size n, the weighted probabilities.}

\item{iterations}{The number of iterations/points.}

\item{pixels}{The number of pixels for the 3D plot.}

\item{width}{Width of the window.}

\item{height}{Height of the window.}

\item{set_palette}{0 = default palette; 1 = grayscale; 2 = rich.colors.}
}
\value{
A list with a 3D array of pixels and the parameters as a JSON.
}
\description{
This function allows you to plot some fractals using the Iterated Function Systems theory.
}
