% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{RunJSONIFS3D}
\alias{RunJSONIFS3D}
\title{RunJSONIFS3D}
\usage{
RunJSONIFS3D(JSON, width, height, set_palette = 0)
}
\arguments{
\item{JSON}{A JSON string, URL or file.}

\item{width}{Width of the window.}

\item{height}{Height of the window.}

\item{set_palette}{0 = default palette; 1 = grayscale; 2 = rich.colors.}
}
\value{
A list with a 3D array of pixels and the parameters as a JSON.
}
\description{
This function allows you to plot some fractals using the Iterated Function Systems theory.
}
\examples{
library(IFSPlot)

url_ifs = "https://git.io/vpSq4"
#result = IFSPlot::RunJSONIFS3D(url_ifs, 866, 866, 0)
}
