% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{RunIFS}
\alias{RunIFS}
\title{RunChaosGame}
\usage{
RunIFS(transformation, translation, probability, iterations, x_res, y_res)
}
\arguments{
\item{transformation}{A list of size n, the 2 x 2 transformation matrices.}

\item{translation}{A list of size n, the 2 x 2 translational vectors.}

\item{probability}{A list of size n, the weighted probabilities.}

\item{iterations}{The number of iterations/points.}
}
\value{
A matrix of black/white pixels.
}
\description{
This function allows you to plot some fractals using the Iterated Function Systems theory.
}
\examples{
pixel_matrix = IFSPlot::RunChaosGame(5, .4, 100**3)
}
