% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{RunIFS2D}
\alias{RunIFS2D}
\title{RunIFS2D}
\usage{
RunIFS2D(transformation, translation, probability, iterations, pixels,
  set_palette = 0, filename = "")
}
\arguments{
\item{transformation}{A list of size n, the 2 x 2 transformation matrices.}

\item{translation}{A list of size n, the 2 x 1 translational vectors.}

\item{probability}{A list of size n, the weighted probabilities.}

\item{iterations}{The number of iterations/points.}

\item{pixels}{The number of pixels for the 2D plot.}

\item{set_palette}{0 = default palette; 1 = grayscale; 2 = rich.colors.}

\item{filename}{If given, saves the image as a png file in the current working directory.}
}
\value{
A list with a 2D array of pixels and the parameters as a JSON.
}
\description{
This function allows you to plot some fractals using the Iterated Function Systems theory.
}
